@@ -1,13 +1,16 @@
 import 'package:Wishy/components/default_button.dart';
 import 'package:Wishy/components/delivery_availability_dialog.dart';
+import 'package:Wishy/components/search_contact.dart';
 import 'package:Wishy/components/search_user.dart';
 import 'package:Wishy/components/variants/variants_widget.dart';
 import 'package:Wishy/constants.dart';
 import 'package:Wishy/global_manager.dart';
 import 'package:Wishy/models/Product.dart';
+import 'package:Wishy/screens/sign_in/sign_in_screen.dart';
 import 'package:Wishy/services/graphql_service.dart';
 import 'package:Wishy/size_config.dart';
 import 'package:Wishy/utils/analytics.dart';
+import 'package:Wishy/utils/is_variants_exists.dart';
 import 'package:flutter/material.dart';
 import 'dart:async';
 
@@ -22,6 +25,7 @@ class RequestData {
 class VariantsAndRequestModal extends StatefulWidget {
   final int productId;
   final String situation, productTitle;
+  final String? cursor;
   final List<Variant> variants;
 
   VariantsAndRequestModal({
@@ -29,6 +33,7 @@ class VariantsAndRequestModal extends StatefulWidget {
     required this.productTitle,
     required this.variants,
     required this.situation,
+    this.cursor,
   });
 
   @override
@@ -43,7 +48,7 @@ class _VariantsAndRequestModalState extends State<VariantsAndRequestModal> {
   RequestData requestData = RequestData();
   Completer<bool>? _phoneValidationCompleter;
 
-  void _onUserSelected(int? userId) {
+  void _onUserSelected(int? userId, bool? isActiveUser) {
     requestData.userId = userId;
   }
 
@@ -53,7 +58,9 @@ class _VariantsAndRequestModalState extends State<VariantsAndRequestModal> {
 
   void _onPhoneChanged(String? phone) {
     requestData.phone = phone;
-    _phoneValidationCompleter!.complete(true);
+    if (_phoneValidationCompleter != null &&
+        !_phoneValidationCompleter!.isCompleted)
+      _phoneValidationCompleter!.complete(true);
   }
 
   void _onReasonChanged(String? reason) {
@@ -72,15 +79,15 @@ class _VariantsAndRequestModalState extends State<VariantsAndRequestModal> {
 
   void _onSubmit() async {
     if (_formKey.currentState!.validate()) {
-      _phoneValidationCompleter = Completer<bool>();
       _formKey.currentState!.save();
-      await _phoneValidationCompleter!.future;
+
+      if (requestData.userId == null) {
+        _phoneValidationCompleter = Completer<bool>();
+        await _phoneValidationCompleter!.future;
+      }
       _onVariantChosen();
 
-      if (requestData.reason != null &&
-              requestData.selectedVariant != null &&
-              requestData.userId != null ||
-          (requestData.name != null && requestData.phone != null)) {
+      if (requestData.selectedVariant != null && requestData.phone != null) {
         try {
           await graphQLQueryHandler("requestProduct", {
             "product_id": widget.productId,
@@ -88,8 +95,18 @@ class _VariantsAndRequestModalState extends State<VariantsAndRequestModal> {
             "phone_number": requestData.phone,
             "reason": requestData.reason,
             "name": requestData.name,
-            "recipient_id": requestData.userId
+            "recipient_id": requestData.userId,
+            "cursor": widget.cursor
           });
+
+          AnalyticsService.trackEvent(analyticEvents["REQUEST_VARIANT_PICKED"]!,
+              properties: {
+                "Product Id": widget.productId,
+                "Variant Id": requestData.selectedVariant!.id,
+                "Reason": requestData.reason,
+                "Name": requestData.name,
+                "Recipient Id": requestData.userId,
+              });
 
           AnalyticsService.trackEvent(analyticEvents["PRODUCT_REQUESTED"]!,
               properties: {
@@ -98,10 +115,11 @@ class _VariantsAndRequestModalState extends State<VariantsAndRequestModal> {
                 "Reason": requestData.reason,
                 "Name": requestData.name,
                 "Recipient Id": requestData.userId,
-                "Variants Exist": widget.variants.length > 1
+                "Variants Exist": isVariantsExists(widget.variants)
               });
           ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-            content: Text("Request sent successfully"),
+            content: Text(
+                "Wish whispered successfully! A notification will be sent when it's noticed."),
           ));
         } catch (error) {
           print(error);
@@ -117,77 +135,80 @@ class _VariantsAndRequestModalState extends State<VariantsAndRequestModal> {
   @override
   Widget build(BuildContext context) {
     if (widget.variants.length > 1) {
-      return Stepper(
-        currentStep: _currentStep,
-        onStepContinue: () {
-          if (_currentStep < 1) {
-            setState(() {
-              _currentStep += 1;
-            });
-            AnalyticsService.trackEvent(
-                analyticEvents["REQUEST_VARIANT_PICKED"]!,
-                properties: {
-                  "Product Id": widget.productId,
-                  "Variant Id": requestData.selectedVariant!.id,
-                  "Reason": requestData.reason,
-                  "Name": requestData.name,
-                  "Recipient Id": requestData.userId,
-                });
-          } else {
-            _onSubmit();
-          }
-        },
-        onStepCancel: () {
-          if (_currentStep > 0) {
-            setState(() {
-              _currentStep -= 1;
-            });
-          } else {
-            Navigator.pop(context);
-          }
-        },
-        steps: [
-          Step(
-            title: Text('Select Your Preferred Variant'),
-            content: Column(
-              children: [
-                VariantsWidget(
-                    productVariants: widget.variants,
-                    withBuyButton: false,
-                    situation: widget.situation,
-                    onVariantChange: _onVariantChange),
-              ],
+      return Column(mainAxisSize: MainAxisSize.min, children: [
+        SizedBox(
+          height: getProportionateScreenHeight(10),
+        ),
+        Text(
+          "Make a Wish Known",
+          style: TextStyle(fontSize: 15, fontWeight: FontWeight.bold),
+        ),
+        Stepper(
+          currentStep: _currentStep,
+          onStepContinue: () {
+            if (_currentStep < 1) {
+              setState(() {
+                _currentStep += 1;
+              });
+            } else {
+              _onSubmit();
+            }
+          },
+          onStepCancel: () {
+            if (_currentStep > 0) {
+              setState(() {
+                _currentStep -= 1;
+              });
+            } else {
+              Navigator.pop(context);
+            }
+          },
+          steps: [
+            Step(
+              title: Text(
+                  'Select Your Wish: Choose preferences that define your ideal item.'),
+              content: Column(
+                children: [
+                  VariantsWidget(
+                      productVariants: widget.variants,
+                      withBuyButton: false,
+                      situation: widget.situation,
+                      onVariantChange: _onVariantChange),
+                ],
+              ),
             ),
-          ),
-          Step(
-            title: Text('Select Whom You\'d Like to Request From'),
-            content: Form(
-                key: _formKey,
-                child: Column(children: [
-                  SearchUserWidget(
-                    onUserSelected: _onUserSelected,
-                    onNameChanged: _onNameChanged,
-                    onPhoneChanged: _onPhoneChanged,
-                  ),
-                  SizedBox(
-                    height: getProportionateScreenHeight(20),
-                  ),
-                  TextFormField(
-                    autovalidateMode: AutovalidateMode.onUserInteraction,
-                    onSaved: _onReasonChanged,
-                    validator: (value) {
-                      if (value!.isEmpty) {
-                        return kReasonNullError;
-                      }
-                      return null;
-                    },
-                    decoration: InputDecoration(labelText: 'Request Reason'),
-                    maxLines: 3,
-                  ),
-                ])),
-          ),
-        ],
-      );
+            Step(
+              title: Text(
+                  "Send a Hint: Let friends know your desired item with a subtle suggestion."),
+              content: Form(
+                  key: _formKey,
+                  child: Column(children: [
+                    SearchContactWidget(
+                      onUserSelected: _onUserSelected,
+                      onNameChanged: _onNameChanged,
+                      onPhoneChanged: _onPhoneChanged,
+                    ),
+                    SizedBox(
+                      height: getProportionateScreenHeight(20),
+                    ),
+                    TextFormField(
+                      autovalidateMode: AutovalidateMode.onUserInteraction,
+                      onSaved: _onReasonChanged,
+                      validator: (value) {
+                        return null;
+                      },
+                      decoration: InputDecoration(
+                          labelText: 'Request Reason',
+                          labelStyle: TextStyle(color: Colors.black),
+                          hintText:
+                              "I found the perfect thing to make our movie nights even better!"),
+                      maxLines: 3,
+                    ),
+                  ])),
+            ),
+          ],
+        )
+      ]);
     } else {
       return Container(
           height: MediaQuery.of(context).size.height * 0.6,
@@ -196,11 +217,20 @@ class _VariantsAndRequestModalState extends State<VariantsAndRequestModal> {
               child: Form(
                   key: _formKey,
                   child: Column(children: [
-                    Text('Select Whom You\'d Like to Request From'),
+                    Text(
+                      "Make a Wish Known",
+                      style:
+                          TextStyle(fontSize: 15, fontWeight: FontWeight.bold),
+                    ),
+                    SizedBox(
+                      height: getProportionateScreenHeight(10),
+                    ),
+                    Text(
+                        "Send a Hint: Let friends know your desired item with a subtle suggestion."),
                     SizedBox(
                       height: getProportionateScreenHeight(20),
                     ),
-                    SearchUserWidget(
+                    SearchContactWidget(
                       onUserSelected: _onUserSelected,
                       onNameChanged: _onNameChanged,
                       onPhoneChanged: _onPhoneChanged,
@@ -240,11 +270,18 @@ class _VariantsAndRequestModalState extends State<VariantsAndRequestModal> {
 }
 
 void showRequestModal(BuildContext context, int productId, String productTitle,
-    List<Variant> variants, String situation) async {
-  if (!GlobalManager().isDeliveryAvailable!) {
+    List<Variant> variants, String situation,
+    {String? cursor = null}) async {
+  if (GlobalManager().signedIn != true) {
+    GlobalManager().setSignInRelatedProductId(productId);
+    Navigator.pushReplacementNamed(context, SignInScreen.routeName);
+    return;
+  }
+
+  if (GlobalManager().isDeliveryAvailable != true) {
     await DeliveryAvailabilityDialog.show(context);
 
-    if (!GlobalManager().isDeliveryAvailable!) return;
+    if (GlobalManager().isDeliveryAvailable != true) return;
   }
 
   showModalBottomSheet<void>(
@@ -265,6 +302,7 @@ void showRequestModal(BuildContext context, int productId, String productTitle,
               situation: situation,
               productTitle: productTitle,
               variants: variants,
+              cursor: cursor,
             ));
       });
 }
