@@ -1,3 +1,6 @@
+import 'package:Wishy/global_manager.dart';
+import 'package:Wishy/screens/complete_profile/complete_profile_screen.dart';
+import 'package:Wishy/screens/sign_in/sign_in_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:Wishy/models/Product.dart';
 
@@ -8,12 +11,29 @@ class CheckoutScreen extends StatelessWidget {
   final Variant variant;
   final int productId;
   final int? recipientId;
+  final String? cursor;
 
   CheckoutScreen(
-      {required this.variant, required this.productId, this.recipientId});
+      {required this.variant,
+      required this.productId,
+      this.recipientId,
+      this.cursor = null});
+
+  void _redirectToSignInIfNeeded(BuildContext context) {
+    if (!GlobalManager().signedIn) {
+      GlobalManager().setSignInRelatedProductId(productId);
+      Navigator.pushReplacementNamed(context, SignInScreen.routeName);
+    } else if (GlobalManager().profileCompleted != true) {
+      GlobalManager().setSignInRelatedProductId(productId);
+      Navigator.pushReplacementNamed(context, CompleteProfileScreen.routeName);
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
+    WidgetsBinding.instance
+        .addPostFrameCallback((_) => _redirectToSignInIfNeeded(context));
+
     return Scaffold(
       appBar: AppBar(
         title: Text('Checkout'),
