@@ -1,4 +1,6 @@
 import 'package:Wishy/utils/contacts.dart';
+import 'package:Wishy/utils/notification.dart';
+import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
 import 'package:Wishy/components/custom_surfix_icon.dart';
 import 'package:Wishy/components/default_button.dart';
@@ -18,30 +20,47 @@ class CompleteProfileForm extends StatefulWidget {
 
 class _CompleteProfileFormState extends State<CompleteProfileForm> {
   final _formKey = GlobalKey<FormState>();
-  String? fullName;
-  String? email;
-  bool _givePermission = true;
+  String? _fullName;
+  String? _email;
+  bool _giveNotificationPermission = true;
 
   void onSubmit() async {
     if (_formKey.currentState!.validate()) {
       _formKey.currentState!.save();
 
+      final email = _email?.trim();
+      final fullName = _fullName?.trim();
       try {
-        if (_givePermission) {
-          await fetchContacts();
+        Map<String, dynamic> notificationParams = {};
+        if (_giveNotificationPermission &&
+            GlobalManager().notificationAvailable == null) {
+          final result = await requestNotificationPermission();
+
+          notificationParams = {
+            "notification_available": result["available"],
+            "fcm_token": result["fcmToken"]
+          };
         }
 
-        await graphQLQueryHandler("updateUserById",
-            {"email": email, "name": fullName, "id": GlobalManager().userId});
-        // await GlobalManager()
-        //     .setParams(newProfileCompleted: true, newUsername: fullName);
+        await graphQLQueryHandler("updateUserById", {
+          "email": email,
+          "name": fullName,
+          "id": GlobalManager().userId,
+          ...notificationParams
+        });
+        GlobalManager().setShowUpAnimation(true);
+        await GlobalManager().setParams(
+            newProfileCompleted: true,
+            newUsername: fullName,
+            newNotificationAvailable:
+                notificationParams["notification_available"]);
         AnalyticsService.trackEvent(
             analyticEvents["COMPLETE_PROFILE_SUBMITTED"]!);
 
         AnalyticsService.setUserProfile(GlobalManager().userId!, {
           "Email": email,
           "Name": fullName,
-          "Contacts Permission": _givePermission
+          "Notification Permission": _giveNotificationPermission
         });
         Navigator.pushNamed(context, LoginSuccessScreen.routeName);
       } catch (error) {
@@ -62,22 +81,42 @@ class _CompleteProfileFormState extends State<CompleteProfileForm> {
           buildFullNameFormField(),
           SizedBox(height: getProportionateScreenHeight(30)),
           buildEmailFormField(),
-          SizedBox(height: getProportionateScreenHeight(20)),
-          CheckboxListTile(
-            title: Text(
-              "Give permission to get contacts for matching with them",
-              style: TextStyle(fontSize: 12),
+          SizedBox(
+              height: getProportionateScreenHeight(
+                  GlobalManager().notificationAvailable == true ? 80 : 50)),
+          // CheckboxListTile(
+          //   title: Text(
+          //     "Give permission to get contacts for matching with them",
+          //     style: TextStyle(fontSize: 12),
+          //   ),
+          //   value: _givePermission,
+          //   onChanged: (bool? value) {
+          //     setState(() {
+          //       _givePermission = value!;
+          //     });
+          //   },
+          //   controlAffinity: ListTileControlAffinity.leading,
+          // ),
+          if (GlobalManager().notificationAvailable == null) ...[
+            CheckboxListTile(
+              title: Text(
+                "Get Personalized Alerts: Receive notifications for customized product recommendations",
+                style: TextStyle(fontSize: 12),
+              ),
+              value: _giveNotificationPermission,
+              onChanged: (bool? value) {
+                setState(() {
+                  _giveNotificationPermission = value!;
+                });
+              },
+              controlAffinity: ListTileControlAffinity.leading,
             ),
-            value: _givePermission,
-            onChanged: (bool? value) {
-              setState(() {
-                _givePermission = value!;
-              });
-            },
-            controlAffinity: ListTileControlAffinity.leading,
+            SizedBox(height: getProportionateScreenHeight(10)),
+          ],
+          DefaultButton(
+            text: "continue",
+            press: onSubmit,
           ),
-          SizedBox(height: getProportionateScreenHeight(40)),
-          DefaultButton(text: "continue", press: onSubmit),
         ],
       ),
     );
@@ -123,7 +162,7 @@ class _CompleteProfileFormState extends State<CompleteProfileForm> {
 
   TextFormField buildFullNameFormField() {
     return TextFormField(
-      onSaved: (newValue) => fullName = newValue,
+      onSaved: (newValue) => _fullName = newValue,
       autovalidateMode: AutovalidateMode.onUserInteraction,
       validator: (value) {
         if (value!.isEmpty) {
@@ -142,10 +181,10 @@ class _CompleteProfileFormState extends State<CompleteProfileForm> {
 
   TextFormField buildEmailFormField() {
     return TextFormField(
-      onSaved: (newValue) => email = newValue,
+      onSaved: (newValue) => _email = newValue,
       autovalidateMode: AutovalidateMode.onUserInteraction,
       validator: (value) {
-        if (value!.isEmpty || !EmailValidator.validate(value)) {
+        if (value!.isEmpty || !EmailValidator.validate(value.trim())) {
           return kInvalidEmailError;
         }
         return null;
