@@ -5,6 +5,7 @@ import 'package:Wishy/global_manager.dart';
 import 'package:Wishy/size_config.dart';
 import 'package:Wishy/services/opt_services.dart';
 import 'package:otp_text_field/otp_field.dart';
+import 'package:otp_text_field/otp_field_style.dart';
 import 'package:otp_text_field/style.dart';
 import 'package:Wishy/utils/analytics.dart';
 import 'package:Wishy/utils/router_utils.dart';
@@ -22,8 +23,10 @@ class OtpForm extends StatefulWidget {
 }
 
 class _OtpFormState extends State<OtpForm> {
-  final otpServices = OTPServices();
+  final otpServices = AuthServices();
   String otpValue = "";
+  bool? completedProfile = GlobalManager().profileCompleted;
+  bool _isPressed = false;
 
   @override
   Widget build(BuildContext context) {
@@ -32,10 +35,12 @@ class _OtpFormState extends State<OtpForm> {
         children: [
           SizedBox(height: SizeConfig.screenHeight * 0.15),
           OTPTextField(
+            onChanged: (v) => {},
             length: 4,
             width: MediaQuery.of(context).size.width,
             fieldWidth: 50,
-            style: TextStyle(fontSize: 17),
+            otpFieldStyle: OtpFieldStyle(focusBorderColor: kPrimaryColor),
+            style: TextStyle(fontSize: 17, color: kPrimaryColor),
             textFieldAlignment: MainAxisAlignment.spaceAround,
             fieldStyle: FieldStyle.box,
             onCompleted: (pin) async {
@@ -56,29 +61,45 @@ class _OtpFormState extends State<OtpForm> {
   }
 
   Future<void> handleOtpSubmission() async {
+    if (_isPressed) return;
+
+    setState(() {
+      _isPressed = true;
+    });
     try {
       final result =
           await otpServices.verifyOTPService(widget.phoneNumber, otpValue);
 
       await GlobalManager().setParams(
-          newToken: result.token,
-          newUserId: result.userId,
-          newUsername: result.username);
+        newToken: result.token,
+        newUserId: result.userId,
+        newUsername: result.username,
+        newProfileCompleted: result.profileCompleted,
+        newNotificationAvailable: result.notificationAvailable,
+        newSignedIn: true,
+      );
+
+      if (result.profileCompleted) {
+        GlobalManager().setShowAnimation(false);
+      }
+
       AnalyticsService.registerSuperProperties({"User Id": result.userId});
+      AnalyticsService.setUserProfile(
+          GlobalManager().userId!, {"Notification Available": false});
 
       AnalyticsService.trackEvent(
         analyticEvents["OPT_SUBMITTED"]!,
       );
 
       if (mounted) {
-        //TODO fix this if that start to use
-        // RouterUtils.routeToHomePage(
-        //     context, result.profileCompleted, result.token);
+        RouterUtils.routeToHomePage();
       }
     } catch (error) {
       ScaffoldMessenger.of(context).showSnackBar(
         SnackBar(content: Text('Error verifying OTP. Please try again.')),
       );
     }
+
+    _isPressed = false;
   }
 }
