@@ -6,12 +6,10 @@ import 'components/body.dart';
 
 class ProfileScreen extends StatelessWidget {
   static String routeName = "/profile";
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: Text("Profile"),
-      ),
       body: Body(),
       bottomNavigationBar: CustomBottomNavBar(selectedMenu: MenuState.profile),
     );
