@@ -1,3 +1,4 @@
+import 'package:Wishy/global_manager.dart';
 import 'package:flutter/material.dart';
 import '../../../size_config.dart';
 import 'sign_form.dart';
@@ -5,30 +6,37 @@ import 'sign_form.dart';
 class Body extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return SingleChildScrollView(
+    return Center(
+        child: SingleChildScrollView(
       child: Padding(
-        padding:
-            EdgeInsets.symmetric(horizontal: getProportionateScreenWidth(20)),
+        padding: EdgeInsets.all(getProportionateScreenWidth(20)),
         child: Column(
           children: [
-            SizedBox(height: SizeConfig.screenHeight * 0.04),
+            CircleAvatar(
+              backgroundColor: Colors.black,
+              radius: getProportionateScreenWidth(65),
+              backgroundImage: AssetImage('assets/images/android_icon.png'),
+            ),
+            SizedBox(height: getProportionateScreenHeight(15)),
             Text(
-              "Welcome",
+              "Welcome to Wishy",
               style: TextStyle(
                 fontSize: getProportionateScreenWidth(28),
                 fontWeight: FontWeight.bold,
               ),
             ),
             Text(
-              "Sign in to your Wishy profile",
+              GlobalManager().signInRelatedProductId == null
+                  ? "Unlock the magic of gifting! Sign up to curate your wishlist and surprise friends with the perfect gifts."
+                  : "Please add your phone number to finalize your action",
               textAlign: TextAlign.center,
             ),
-            SizedBox(height: SizeConfig.screenHeight * 0.08),
+            SizedBox(height: getProportionateScreenHeight(30)),
             SignForm(),
-            SizedBox(height: SizeConfig.screenHeight * 0.08),
+            SizedBox(height: getProportionateScreenHeight(20)),
           ],
         ),
       ),
-    );
+    ));
   }
 }
